NList.mui: 64 bit and compiler lint fixes

git-svn-id: https://svn.aros.org/svn/aros/trunk@37247 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/mui/classes/nlist/nlist_mcc/NList_mcc1.c
@@ -996,10 +996,10 @@ IPTR mNL_Set(struct IClass *cl,Object *obj,Msg msg)
         REDRAW;
         break;
       case MUIA_NList_MakeActive :
-        data->NList_MakeActive = (LONG) tag->ti_Data;
+        data->NList_MakeActive = tag->ti_Data;
         break;
       case MUIA_NList_KeepActive :
-        data->NList_KeepActive = (LONG) tag->ti_Data;
+        data->NList_KeepActive = tag->ti_Data;
         break;
       case MUIA_NList_TypeSelect :
         MOREQUIET;
